+++ src/visualize/+mag/+graphics/+view/Spectrogram.m
@@ -1,5 +1,5 @@
-classdef Frequency < mag.graphics.view.View
-% FREQUENCY Show PSD and spectrogram of magnetic field.
+classdef Spectrogram < mag.graphics.view.View
+% SPECTROGRAM Show spectrogram of magnetic field.
 
     properties
         % NORMALIZE Normalize data before computing spectrum to highlight
@@ -13,19 +13,15 @@ classdef Frequency < mag.graphics.view.View
         Window (1, 1) double = missing()
         % OVERLAP Number of overlapped samples.
         Overlap (1, 1) double = missing()
-        % PSDSTART Start date of PSD plot.
-        PSDStart (1, 1) datetime = NaT(TimeZone = "UTC")
-        % PSDDURATION Duration of PSD plot.
-        PSDDuration (1, 1) duration = hours(1)
     end
 
     methods
 
-        function this = Frequency(results, options)
+        function this = Spectrogram(results, options)
 
             arguments
                 results
-                options.?mag.graphics.view.Frequency
+                options.?mag.graphics.view.Spectrogram
             end
 
             this.Results = results;
@@ -51,7 +47,7 @@ classdef Frequency < mag.graphics.view.View
             primaryCharts = this.getFrequencyCharts(primary, primarySpectrum, primarySensor, "left");
             secondaryCharts = this.getFrequencyCharts(secondary, secondarySpectrum, secondarySensor, "right");
 
-            this.Figures(1) = this.Factory.assemble( ...
+            this.Figures = this.Factory.assemble( ...
                 primaryCharts{:}, ...
                 secondaryCharts{:}, ...
                 Title = this.getFrequencyFigureTitle(primary, secondary), ...
@@ -59,32 +55,6 @@ classdef Frequency < mag.graphics.view.View
                 Arrangement = [9, 2], ...
                 LinkXAxes = true, ...
                 TileIndexing = "columnmajor", ...
-                WindowState = "maximized");
-
-            % PSD and spectrogram.
-            if ismissing(this.PSDStart) || ~isbetween(this.PSDStart, primary.Time(1), primary.Time(end))
-                psdStart = primary.Time(1);
-            else
-                psdStart = this.PSDStart;
-            end
-
-            if (this.PSDDuration > (primary.Time(end) - psdStart))
-                psdDuration = primary.Time(end) - psdStart;
-            else
-                psdDuration = this.PSDDuration;
-            end
-
-            psdPrimary = primary.computePSD(Start = psdStart, Duration = psdDuration);
-            psdSecondary = secondary.computePSD(Start = psdStart, Duration = psdDuration);
-
-            yLine = mag.graphics.chart.Line(Axis = "y", Value = 0.01, Style = "--", Label = "10 pT Hz^{-0.5}");
-
-            this.Figures(2) = this.Factory.assemble( ...
-                psdPrimary, mag.graphics.style.Default(Title = compose("%s PSD", primarySensor), XLabel = this.FLabel, YLabel = this.PSDLabel, XScale = "log", YScale = "log", Legend = ["x", "y", "z"], Charts = [mag.graphics.chart.Plot(XVariable = "Frequency", YVariables = ["X", "Y", "Z"]), yLine]), ...
-                psdSecondary, mag.graphics.style.Default(Title = compose("%s PSD", secondarySensor), XLabel = this.FLabel, YLabel = this.PSDLabel, XScale = "log", YScale = "log", Legend = ["x", "y", "z"], Charts = [mag.graphics.chart.Plot(XVariable = "Frequency", YVariables = ["X", "Y", "Z"]), yLine]), ...
-                Title = this.getPSDFigureTitle(primary, secondary, psdStart, psdDuration), ...
-                Name = this.getPSDFigureName(primary, secondary, psdStart), ...
-                Arrangement = [2, 1], ...
                 WindowState = "maximized");
         end
     end
@@ -108,14 +78,6 @@ classdef Frequency < mag.graphics.view.View
 
         function value = getFrequencyFigureName(this, primary, secondary)
             value = this.getFrequencyFigureTitle(primary, secondary) + compose(" Frequency (%s)", this.date2str(primary.MetaData.Timestamp));
-        end
-
-        function value = getPSDFigureTitle(this, primary, secondary, psdStart, psdDuration)
-            value = compose("Start: %s - Duration: %s - (%d, %d)", this.date2str(psdStart), psdDuration, primary.MetaData.getDisplay("DataFrequency"), secondary.MetaData.getDisplay("DataFrequency"));
-        end
-
-        function value = getPSDFigureName(this, primary, secondary, psdStart)
-            value = compose("%s (%d, %d) PSD (%s)", primary.MetaData.getDisplay("Mode"), primary.MetaData.getDisplay("DataFrequency"), secondary.MetaData.getDisplay("DataFrequency"), this.date2str(psdStart));
         end
     end
 end